remove an unused method; trivial type changes

+++ macosx/CTGradient/CTGradientAdditions.h
@@ -26,7 +26,7 @@
 
 @interface CTGradient (ProgressBar)
 
-+ (CTGradient *) progressGradientForRed: (float) redComponent green: (float) greenComponent blue: (float) blueComponent;
++ (CTGradient *) progressGradientForRed: (CGFloat) redComponent green: (CGFloat) greenComponent blue: (CGFloat) blueComponent;
 
 + (CTGradient *)progressWhiteGradient;
 + (CTGradient *)progressGrayGradient;
